Working on some Security Flaws
In the rush to deploy a MVP prior to our presentation the front end code was left with a few security flaws. We have exposed our API keys to the public and that really needs to be addressed. Working on the backend at the same time so the keys can be stored and accessed from their when the user has valid credentials. Cleaned up some of the code but found a major bug on the agenda items page that crashes the site locally.